# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/login":{"post":{"operationId":"LoginController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["login"]}},"/api/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["users"]},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["users"]}},"/api/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["users"]},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["users"]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["users"]}},"/api/roles":{"post":{"operationId":"RolesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleDto"}}}},"responses":{"201":{"description":""}},"tags":["roles"]},"get":{"operationId":"RolesController_filterQuery","parameters":[],"responses":{"200":{"description":""}},"tags":["roles"]}},"/api/roles/all":{"get":{"operationId":"RolesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["roles"]}},"/api/roles/{id}":{"patch":{"operationId":"RolesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleDto"}}}},"responses":{"200":{"description":""}},"tags":["roles"]},"delete":{"operationId":"RolesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["roles"]}},"/api/upload":{"post":{"operationId":"UploadController_uploadFile","parameters":[],"responses":{"201":{"description":""}},"tags":["upload"]}},"/api/compon":{"post":{"operationId":"ComponController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateComponDto"}}}},"responses":{"201":{"description":""}},"tags":["compons"]},"get":{"operationId":"ComponController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["compons"]}},"/api/compon/{id}":{"patch":{"operationId":"ComponController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateComponDto"}}}},"responses":{"200":{"description":""}},"tags":["compons"]},"delete":{"operationId":"ComponController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateComponDto"}}}},"responses":{"200":{"description":""}},"tags":["compons"]}},"/api/sms":{"post":{"operationId":"SmsController_sendSms","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSmsDto"}}}},"responses":{"201":{"description":""}},"tags":["sms"]}},"/category":{"post":{"operationId":"CategoryController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"CategoryController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/category/{id}":{"get":{"operationId":"CategoryController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"CategoryController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoryDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"CategoryController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/attr-key":{"post":{"operationId":"AttrKeyController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAttrKeyDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"AttrKeyController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/attr-key/{id}":{"get":{"operationId":"AttrKeyController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"AttrKeyController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAttrKeyDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"AttrKeyController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/attr-val":{"post":{"operationId":"AttrValController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAttrValDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"AttrValController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/attr-val/{id}":{"get":{"operationId":"AttrValController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"AttrValController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAttrValDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"AttrValController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/product":{"post":{"operationId":"ProductController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"ProductController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/product/{id}":{"get":{"operationId":"ProductController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"ProductController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"ProductController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/product-specs":{"post":{"operationId":"ProductSpecsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductSpecDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"ProductSpecsController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/product-specs/{id}":{"get":{"operationId":"ProductSpecsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"ProductSpecsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductSpecDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"ProductSpecsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/brand":{"post":{"operationId":"BrandController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBrandDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"BrandController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/brand/{id}":{"get":{"operationId":"BrandController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"BrandController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBrandDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"BrandController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"JinPiKa","description":"nest-admin后台接口","version":"1.0","contact":{"name":"JinPiKa","url":"https://gitee.com/jinxin0517/nest-admin","email":"847164495@qq.com"}},"tags":[],"servers":[],"components":{"schemas":{"LoginDto":{"type":"object","properties":{"username":{"type":"string","uniqueItems":true},"password":{"type":"string"},"loginType":{"type":"string","enum":["phone","account"]}},"required":["username","password","loginType"]},"CreateUserDto":{"type":"object","properties":{"username":{"type":"string","uniqueItems":true},"email":{"type":"string","uniqueItems":true},"phone":{"type":"string","uniqueItems":true},"role":{"type":"array","items":{"type":"string"}},"avatar":{"type":"string"},"password":{"type":"string"},"captcha":{"type":"string"}},"required":["username","email","phone","role","avatar","password","captcha"]},"UpdateUserDto":{"type":"object","properties":{}},"CreateRoleDto":{"type":"object","properties":{"name":{"type":"string","uniqueItems":true},"desc":{"type":"string"},"compon":{"type":"array","items":{"type":"string"}},"half_compon":{"type":"array","items":{"type":"string"}}},"required":["name","desc","compon","half_compon"]},"UpdateRoleDto":{"type":"object","properties":{"name":{"type":"string","uniqueItems":true},"desc":{"type":"string"},"compon":{"type":"array","items":{"type":"string"}},"half_compon":{"type":"array","items":{"type":"string"}}}},"CreateComponDto":{"type":"object","properties":{"name":{"type":"string","uniqueItems":true},"type":{"type":"string"},"parentId":{"type":"string"}},"required":["name","type","parentId"]},"UpdateComponDto":{"type":"object","properties":{"name":{"type":"string","uniqueItems":true},"type":{"type":"string"},"parentId":{"type":"string"}}},"CreateSmsDto":{"type":"object","properties":{"phone":{"type":"string"}},"required":["phone"]},"CreateCategoryDto":{"type":"object","properties":{}},"UpdateCategoryDto":{"type":"object","properties":{}},"CreateAttrKeyDto":{"type":"object","properties":{}},"UpdateAttrKeyDto":{"type":"object","properties":{}},"CreateAttrValDto":{"type":"object","properties":{}},"UpdateAttrValDto":{"type":"object","properties":{}},"CreateProductDto":{"type":"object","properties":{}},"UpdateProductDto":{"type":"object","properties":{}},"CreateProductSpecDto":{"type":"object","properties":{}},"UpdateProductSpecDto":{"type":"object","properties":{}},"CreateBrandDto":{"type":"object","properties":{}},"UpdateBrandDto":{"type":"object","properties":{}}}}}